Gesture Tracking/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astroid==2.5.1
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cx-Oracle==8.1.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.2
dlib==19.22.0
dnspython==1.16.0
email-validator==1.1.2
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imutils==0.5.4
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mongoengine==0.23.1
MouseInfo==0.1.3
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.0.3
neo4j==4.2.1
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.2.4
Pillow==8.2.0
prettytable==2.1.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pylint==2.7.2
pymongo==3.11.4
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PySocks==1.7.1
python-dateutil==2.8.1
python-snappy==0.6.0
PyTweening==1.0.3
pytz==2021.1
pywin32==300
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.18
stem==1.8.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc3
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2021.4.8
toml==0.10.2
torrequest==0.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.7
wcwidth==0.2.5
Werkzeug==2.0.1
winkerberos==0.8.0
wrapt==1.12.1
WTForms==2.3.3
